- Fixed order push endpoint - Renamed a couple of files to fix spelling errors and be consistent with Magento naming conventions.
