+++ Api/etc/di.xml
@@ -22,7 +22,7 @@
         <arguments>
             <argument  name="upcFinder" xsi:type="object">RetailOps\Api\Model\Api\Map\UpcFinder</argument>
             <argument  name="calculateDiscount" xsi:type="object">RetailOps\Api\Service\CalculateOrderDiscount</argument>
-            <argument  name="calculateAmount" xsi:type="object">RetailOps\Api\Service\Order\Map\CalculateAmout</argument>
+            <argument  name="calculateAmount" xsi:type="object">RetailOps\Api\Service\Order\Map\CalculateAmount</argument>
             <argument  name="calculateItemPrice" xsi:type="object">RetailOps\Api\Service\CalculateItemPrice</argument>
         </arguments>
     </type>
@@ -59,7 +59,7 @@
         <arguments>
             <argument name="routerList" xsi:type="array">
                 <item name="retailops" xsi:type="array">
-                    <item name="class" xsi:type="string">\RetailOps\Api\CustomRouter\Router</item>
+                    <item name="class" xsi:type="string">\RetailOps\Api\Controller\Router</item>
                     <item name="disable" xsi:type="boolean">false</item>
                     <item name="sortOrder" xsi:type="string">22</item>
                 </item>
